requirements.txt
fastapi
uvicorn
python-telegram-bot[webhooks]==20.6
python-dotenv==1.0.1
requests